requirements.txt
# Основное для Discord-бота
discord.py>=2.4.0

# Для загрузки .env файлов
python-dotenv>=1.0.1

# Для работы с файлами асинхронно
aiofiles>=23.2.1

# Для работы с базой данных SQLite асинхронно
aiosqlite>=0.20.0

# Для сетевых запросов (если будешь добавлять API-функции)
requests>=2.32.3

# Асинхронность (входит в стандарт Python, но оставим для совместимости)
asyncio
